pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "drt-core"
version = "0.4.3"
description = "Reverse ETL for the code-first data stack"
readme = "README.md"
license = { file = "LICENSE" }
authors = [{ name = "drt-hub", email = "masukai9612kf@gmail.com" }]
keywords = ["reverse-etl", "data-engineering", "dbt", "bigquery", "cli"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Database",
]
requires-python = ">=3.10"
dependencies = [
    "typer>=0.12",
    "pydantic>=2.0",
    "jinja2>=3.1",
    "httpx>=0.27",
    "pyyaml>=6.0",
    "rich>=13.0",
]

[project.optional-dependencies]
bigquery = ["google-cloud-bigquery>=3.0"]
duckdb = ["duckdb>=0.10"]
postgres = ["psycopg2-binary>=2.9"]
redshift = ["psycopg2-binary>=2.9"]  # Redshift uses PostgreSQL wire protocol
clickhouse = ["clickhouse-connect>=0.7.0"]
sheets = ["google-api-python-client>=2.0", "google-auth>=2.0"]
mysql = ["pymysql>=1.1"]
mcp = ["fastmcp>=2.0"]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "pytest-httpserver>=1.0",
    "mypy>=1.10",
    "ruff>=0.4",
    "types-pyyaml",
    "types-PyMySQL",
]

[project.scripts]
drt = "drt.cli.main:app"

[project.urls]
Homepage = "https://github.com/drt-hub/drt"
Repository = "https://github.com/drt-hub/drt"
Issues = "https://github.com/drt-hub/drt/issues"

[tool.hatch.build.targets.wheel]
packages = ["drt"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]

[tool.mypy]
python_version = "3.10"
strict = true
ignore_missing_imports = true
warn_unused_ignores = false

[[tool.mypy.overrides]]
# fastmcp decorators are untyped — skip strict checks for MCP server module
module = "drt.mcp.server"
disallow_untyped_decorators = false

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
